ugc/api/requirements.txt
fastapi
pydantic
aiokafka
orjson
python-jose[cryptography]
uvicorn
gunicorn
sentry-sdk
python-logstash-async
backoff
pymongo[srv]
motor
